% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/points_to_line.R
\name{points_to_line}
\alias{points_to_line}
\title{Function to convert a table of XY coordinates to spatial lines}
\usage{
points_to_line(data, long, lat, id_field = NULL, sort_field = NULL)
}
\description{
Function to convert a table of XY coordinates to spatial lines
}

