% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{quantile_labels}
\alias{quantile_labels}
\title{Create nice-looking quantile labels for Leaflet mapping.}
\usage{
quantile_labels(vec, n)
}
\arguments{
\item{vec}{The column of data that you are visualizing on your choropleth map}

\item{n}{The number of classes you've chosen}
}
\description{
At present, the amazing \code{leaflet} package uses percentiles in its quantile legends; however, sometimes you
 want to show the actual values.  This function allows you to do just that.
}

