% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r2cartodb.R
\name{r2cartodb}
\alias{r2cartodb}
\title{r2cartodb: a function to import data to your CartoDB account.}
\usage{
r2cartodb(obj, layer_name, account_id, api_key)
}
\description{
Set your working directory before calling the function.
Also, the function depends on the rgdal and httr packages, and requires that you have Rtools installed
(and that you have Rtools in your PATH if you are on Windows) if you want to upload spatial
data frames.
}

