% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{transform_xy}
\alias{transform_xy}
\title{Helper functions to perform common geospatial tasks}
\usage{
transform_xy(sp_object)
}
\description{
Helper functions to perform common geospatial tasks
}

