% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wdi_leaflet.R
\name{wdi_leaflet}
\alias{wdi_leaflet}
\title{Function to create a Leaflet interactive map in RStudio from a World Bank indicator.}
\usage{
wdi_leaflet(indicator, indicator_alias = "Value", year = 2012,
  classes = 5, colors = "Blues")
}
\description{
Function to create a Leaflet interactive map in RStudio from a World Bank indicator.
}

