% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{transform_xy}
\alias{transform_xy}
\title{Quick transform of spatial objects to WGS84}
\usage{
transform_xy(sp_object)
}
\arguments{
\item{sp_object}{an R object of class \code{Spatial*} that you'd like to transform to WGS84.}
}
\description{
The function will use \code{spTransform} from the \code{rgdal} package to automatically transform your
 spatial data to the WGS84 geographic coordinate system.  As this is the coordinate system required for mapping
 with the \code{leaflet} package, it can come in handy when mapping spatial objects in R.
}

